requirements.txt
tensorflow
numpy
pillow
tkinter
scipy
matplotlib
scikit-learn